Chore - move files and folders to separate the spec from autospy schematics - each containing its /files folder and its dependencies
