+++ src/spec/files/__specFileName__.template
