+++ tsconfig.json
@@ -23,7 +23,7 @@
   "exclude": [
     "src/**/files/**/*",
     "src/**/*_spec.ts",
-    "src/read/example",
+    "src/spec/read/example",
     "tests",
     "lib/third_party",
     "lib/utility/*_spec.ts",
